requirements.txt
pandas
numpy
scikit-learn
mlflow
pyyaml
fastapi
uvicorn
joblib
pydantic